16-148: task tests
