+++ udemy/complete-nodejs-dev/task-manager/package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "start": "node src/index.js",
     "dev": "env-cmd ./config/dev.env nodemon src/index.js",
-    "test": "env-cmd ./config/test.env jest --watch"
+    "test": "env-cmd ./config/test.env jest --watch --runInBand"
   },
   "jest": {
     "testEnvironment": "node"
